web/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-ldap==3.4.4
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0
python-multipart==0.0.6
pyyaml==6.0.1
